requirements.txt
Flask==2.1.0
gunicorn==20.1.0
google-cloud-language>=2.0,<3
google-cloud-bigquery>=3.0,<4

